⛏ fix: fix jpa bugs

+++ backend/src/main/java/com/auctopus/project/api/service/AuctionService.java
@@ -4,7 +4,6 @@ import com.auctopus.project.api.request.AuctionUpdateRequest;
 import com.auctopus.project.db.domain.Auction;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
 
 /**
  * 경매방 관련 로직 처리를 위한 서비스 구현
@@ -17,11 +16,11 @@ public interface AuctionService {
 
     List<Auction> getAuctionListByStartTime(String word, Pageable pageable);
 
-    List<Auction> getAuctionListByCategorySeq(Long likeCategorySeq, Pageable pageable);
+    List<Auction> getAuctionListByCategorySeq(int likeCategorySeq, Pageable pageable);
 
-    List<Auction> getAuctionListTodayAndCategorySeq(Long categorySeq, Pageable pageable);
+    List<Auction> getAuctionListTodayAndCategorySeq(int categorySeq, Pageable pageable);
 
-    List<Auction> getAllAuctionListByCategorySeq(Long categorySeq, Pageable pageable);
+    List<Auction> getAllAuctionListByCategorySeq(int categorySeq, Pageable pageable);
 
     // 한 개의 경매방 정보 보기
     Auction getAuction(int auctionSeq);
